# Slight store performance improvements (low-hanging fruits)