This:
- removes cloning the external symbols `HashSet` on each symbol internalization from the store (this could probably also be removed from the repl),
- adds a few const functions where possible,
- and a few minor cleanups:
```
go_base_10_16_bls12     time:   [3.7405 ms 3.7457 ms 3.7512 ms]
                        change: [-27.969% -27.813% -27.653%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

go_base_10_160_bls12    time:   [35.656 ms 35.686 ms 35.717 ms]
                        change: [-28.890% -28.219% -27.799%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe

go_base_10_16_pasta_pallas
                        time:   [3.8389 ms 3.8437 ms 3.8493 ms]
                        change: [-28.424% -28.246% -28.066%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

go_base_10_160_pasta_pallas
                        time:   [37.019 ms 37.131 ms 37.249 ms]
                        change: [-28.107% -27.759% -27.413%] (p = 0.00 < 0.05)
                        Performance has improved.
```
See also the before/after flamegraph linked in the first commit.


### Edit: after replacing `package_name` by an `Arc`

```
go_base_10_16_bls12     time:   [3.7013 ms 3.7050 ms 3.7092 ms]
                        change: [-27.409% -27.190% -26.967%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

go_base_10_160_bls12    time:   [35.369 ms 35.423 ms 35.495 ms]
                        change: [-27.675% -27.452% -27.235%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe

go_base_10_16_pasta_pallas
                        time:   [3.8327 ms 3.8370 ms 3.8416 ms]
                        change: [-28.351% -27.917% -27.536%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

go_base_10_160_pasta_pallas
                        time:   [36.550 ms 36.587 ms 36.625 ms]
                        change: [-27.235% -27.099% -26.971%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
```