@huitseeker 2023-03-20
TODO (for myself): use an `Rc` in the package representation, and using that instead of the package name, for extensibility.